# auth/src/swagger/swagger.yml
openapi: 3.0.0
info:
  title: Software2 API Documentation
  description: |
    This is the API documentation for the Software2 project. It provides details about the available authentication, permission, and role endpoints, their usage, and expected inputs/outputs.
  termsOfService: https://example.com/terms/
  contact:
    email: support@software2.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.software2.com
    description: Production server
...
paths:
...
  /auth/verify:
    post:
      tags:
        - auth
      summary: Verify user account
      description: Verify a user's account using a verification code.
      operationId: verifyCode
      requestBody:
        description: Verification data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "john.doe@example.com"
                code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: Account verified successfully
        '400':
          description: Invalid or expired code
        '500':
          description: Server error
  /auth/resend:
    post:
      tags:
        - auth
      summary: Resend verification code
      description: Resend the verification code to the user's email.
      operationId: resendVerificationCode
      requestBody:
        description: Resend verification data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "john.doe@example.com"
      responses:
        '200':
          description: Verification code resent successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
  /auth/2fa:
    post:
      tags:
        - auth
      summary: Two-factor authentication
      description: Authenticate a user using two-factor authentication.
      operationId: secondFactorAuthentication
      requestBody:
        description: Two-factor authentication data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "john.doe@example.com"
                code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: Authentication successful
        '400':
          description: Invalid or expired code
        '500':
          description: Server error
  /auth/changepassword/{id}:
    patch:
      tags:
        - auth
      summary: Change user password
      description: Change the password of a user.
      operationId: changePassword
      parameters:
        - name: 68156310a2bcde0782740346
          in: path
          required: true
          schema:
            type: string
          description: ID of the user
      requestBody:
        description: Password change data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  example: "Correcp@ssword123"
                newPassword:
                  type: string
                  example: "Correcp@ssword1234"
                confirmNewPassword:
                  type: string
                  example: "Correcp@ssword1234"
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
  /auth/resendsms:
    patch:
      tags:
        - auth
      summary: Resend SMS verification code
      description: Resend the SMS verification code to the user's phone.
      operationId: resendSMS
      requestBody:
        description: Resend SMS data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                number:
                  type: string
                  example: "3001234567"
      responses:
        '200':
          description: SMS code resent successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
  /auth/verifypasscode:
    post:
      tags:
        - auth
      summary: Verify password reset code
      description: Verify the password reset code sent to the user.
      operationId: verifyPassCode
      requestBody:
        description: Verification data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "john.doe@example.com"
                code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: Code verified successfully
        '400':
          description: Invalid or expired code
        '500':
          description: Server error
  /auth/resetpassword:
    post:
      tags:
        - auth
      summary: Reset user password
      description: Send a password reset code to the user's email.
      operationId: resetPassword
      requestBody:
        description: Reset password data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "john.doe@example.com"
      responses:
        '200':
          description: Password reset code sent successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
  /auth/changeresetpassword/{id}:
    put:
      tags:
        - auth
      summary: Change reset password
      description: Change the password after verifying the reset code.
      operationId: ChangeResetPassword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the user
      requestBody:
        description: Change reset password data
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  example: "Correcp@ssword12345"
                confirmNewPassword:
                  type: string
                  example: "Correcp@ssword12345"
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
...